+++ semgrep/semgrep/evaluation.py
@@ -20,6 +20,7 @@ from semgrep.semgrep_types import pattern_name_for_operator
 from semgrep.semgrep_types import pattern_names_for_operator
 from semgrep.semgrep_types import PatternId
 from semgrep.semgrep_types import Range
+from semgrep.semgrep_types import TAINT_MODE
 from semgrep.util import debug_print
 from semgrep.util import flatten
 from semgrep.util import print_stderr
@@ -206,16 +207,21 @@ def evaluate(
         }
     ]
     debug_print(str(pattern_ids_to_pattern_matches))
-    valid_ranges_to_output = evaluate_expression(
-        rule.expression,
-        pattern_ids_to_pattern_matches,
-        flags={RCE_RULE_FLAG: allow_exec},
-        steps_for_debugging=steps_for_debugging,
-    )
+    if rule.mode == TAINT_MODE:
+        valid_ranges_to_output = {
+            pattern_match.range for pattern_match in pattern_matches
+        }
+    else:
+        valid_ranges_to_output = evaluate_expression(
+            rule.expression,
+            pattern_ids_to_pattern_matches,
+            flags={RCE_RULE_FLAG: allow_exec},
+            steps_for_debugging=steps_for_debugging,
+        )
 
-    # only output matches which are inside these offsets!
-    debug_print(f"compiled result {valid_ranges_to_output}")
-    debug_print("-" * 80)
+        # only output matches which are inside these offsets!
+        debug_print(f"compiled result {valid_ranges_to_output}")
+        debug_print("-" * 80)
 
     for pattern_match in pattern_matches:
         if pattern_match.range in valid_ranges_to_output: